Fixes https://github.com/honeycombio/honeyaws/issues/64
If honeyelb/honeycloudfront was used with no arguments, it would
try to ingest logs from all load balancers/distributions. However,
if one of them had not logs enabled, it would terminate with a fatal
error. This made it very hard to have a single binary ingesting logs
from all load balancers/distributions in an account, since every time
a new lb/distro was created without logs it would break existing
running daemons.

This commit fixes this by adding a check on the ingest logic. In
case no positional argument is passed to honeyelb/honeycloudfront,
it will NOT be terminated the process if there is an error when
attempting to create a 'downloader' process for a lb/distro. Instead,
a error-level log will be displayed.

This means that it's technically possible for *all* lb/distros to
fail and no logs to be ingested.